move stuff into quizbowl/ folder
