+++ quizbowl/Player.js
@@ -1,11 +1,10 @@
-import { USERNAME_MAX_LENGTH } from './constants.js';
+import { USERNAME_MAX_LENGTH } from '../server/multiplayer/constants.js';
 
 class Player {
   constructor (userId) {
     this.userId = userId;
 
     this.username = '';
-    this.online = true;
 
     this.powers = 0;
     this.tens = 0;
@@ -71,7 +70,7 @@ class Player {
    * @param {number} [maxLength=USERNAME_MAX_LENGTH]
    * @returns {string} newUsername
    */
-  updateUsername (username, maxLength = USERNAME_MAX_LENGTH) {
+  safelySetUsername (username, maxLength = USERNAME_MAX_LENGTH) {
     if (!username) { username = ''; }
     username = username.substring(0, maxLength);
     this.username = username;
