+++ server/multiplayer/ServerPlayer.js
@@ -1,0 +1,8 @@
+import Player from '../../quizbowl/Player.js';
+
+export default class ServerPlayer extends Player {
+  constructor (userId) {
+    super(userId);
+    this.online = true;
+  }
+}
